Remove number from HPA tissues with cell types
Tissue number from tissues with cell type specified were eliminated: skin 1 and 2, soft tissue 1 and 2. The parser ignores the numbers in tissues for which the cell type are specified (e.g. skin 1 hair follicles is searched as skin hair follicles). The parser does not ignore the number in tissues for which cell types are not specified or cell types containing numbers (e.g. Alveolar cells type 1)